chore: refactor reused code
