+++ packages/shared/src/components/sources/UnfeaturedSourceCard.tsx
@@ -2,10 +2,7 @@ import React, { ReactElement } from 'react';
 import classNames from 'classnames';
 import { useRouter } from 'next/router';
 import { Card, CardLink } from '../cards/Card';
-import { SourceType } from '../../graphql/sources';
-import { Button, ButtonVariant } from '../buttons/Button';
-import { SquadJoinButton } from '../squads/SquadJoinButton';
-import { Origin } from '../../lib/log';
+import { ButtonVariant } from '../buttons/Button';
 import {
   Typography,
   TypographyColor,
@@ -14,8 +11,9 @@ import {
 } from '../typography/Typography';
 import { largeNumberFormat } from '../../lib';
 import { Separator } from '../cards/common';
-import { UnFeaturedSourceCardProps } from './commonTypes';
-import { SourceImage } from './SourceImage';
+import { UnFeaturedSourceCardProps } from './common/types';
+import { SourceImage } from './common/SourceImage';
+import { SourceJoinButton } from './common/SourceJoinButton';
 
 export const UnfeaturedSourceCard = ({
   source,
@@ -41,30 +39,14 @@ export const UnfeaturedSourceCard = ({
           title={title}
           size="size-16"
         />
-        {!!action &&
-        action?.type === 'action' &&
-        source?.type === SourceType.Squad ? (
-          <SquadJoinButton
-            className={{ button: '!btn-tertiaryFloat' }}
-            squad={source}
-            origin={Origin.SquadDirectory}
-            onSuccess={() => router.push(source?.permalink)}
-            joinText={action?.text}
-            data-testid="squad-action"
-          />
-        ) : (
-          <Button
-            variant={ButtonVariant.Float}
-            onClick={action?.type === 'action' ? action?.onClick : undefined}
-            tag={action?.type === 'link' ? 'a' : undefined}
-            href={action?.type === 'link' && action.href}
-            target={action?.target ? action.target : '_self'}
-            rel="noopener"
-            data-testid="source-action"
-          >
-            {action?.text}
-          </Button>
-        )}
+        <SourceJoinButton
+          action={action}
+          source={source}
+          variant={ButtonVariant.Float}
+          className={{
+            squadJoinButton: '!btn-tertiaryFloat',
+          }}
+        />
       </div>
       <Typography
         tag={TypographyTag.H1}
