+++ packages/shared/src/components/sources/common/SourceImage.tsx
