+++ packages/shared/src/components/sources/common/types.ts
@@ -1,5 +1,5 @@
 import { ReactNode } from 'react';
-import { Squad } from '../../graphql/sources';
+import { Squad } from '../../../graphql/sources';
 
 type SourceCardActionType = 'link' | 'action';
 export interface SourceCardAction {
